changing the management interface standalone.*xml
for not requiring the http call authentication
the http call is used during readiness probe to check
if eap is ready. If ADMIN_USER/ADMIN_PASSWORD
env variables are not set then the python REST API
goes without authentication. We need to allow
this and changing the standalone-openshift.xml
for this behaviour.